----handle concurrent, lock-----

+++ migrations/20250910162131-,Create_Order_table.js
@@ -33,7 +33,7 @@ module.exports = {
         defaultValue: 'pending',
       },
       total_amount: {
-        type: Sequelize.DECIMAL(12, 2),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       created_at: {
